+++ src/Renderers/RendererContract.php
@@ -15,14 +15,22 @@ declare(strict_types=1);
 namespace MarcinOrlowski\TextTable\Renderers;
 
 use MarcinOrlowski\TextTable\TextTable;
-use MarcinOrlowski\TextTable\Output\WriterContract;
 
 interface RendererContract
 {
     /**
      * Renders provided `Table` using provided output writer.
      *
-     * @param TextTable $table Instance of `AsciiTable` to render.
+     * @param TextTable $table Instance of `TextTable` to render.
      */
     public function render(TextTable $table): array;
+
+    /**
+     * Helper that returns rendered table as single string.
+     *
+     * @param TextTable $table Instance of `TextTable` to render.
+     *
+     * @return string
+     */
+    public function renderAsString(TextTable $table): string;
 }